// src/types.ts
export type Customer = {
  id: string;
  email?: string;
  name?: string;
  browser?: string;
  created_at?: string;
  current_url?: string;
  external_id?: any;
  profile_photo_url?: string;
  first_seen?: any;
  host?: string;
  ip?: string;
  // Datetime
  last_seen_at: string;
  metadata?: any;
  os?: string;
  pathname?: string;
  phone?: string | number;
  tags?: Array<Tag>;
  time_zone?: string;
  updated_at?: string;
  title: string;
  // Associations
  company?: Company;
  conversations?: Array<Conversation>;
};

export type Company = {
  id: string;
  name: string;
  description?: string;
  website_url?: string;
  external_id?: string;
  slack_channel_id?: string;
  slack_channel_name?: string;
  slack_team_id?: string;
  slack_team_name?: string;
  updated_at: string;
};

export type MessageType = 'reply' | 'note' | 'bot';
export type MessageSource =
  | 'chat'
  | 'slack'
  | 'mattermost'
  | 'email'
  | 'sms'
  | 'api'
  | 'sandbox';

export type Message = {
  id: string;
  body: string;
  type?: MessageType;
  source?: MessageSource;
  private?: boolean;
  created_at: string;
  sent_at?: string;
  seen_at?: string;
  customer_id?: string;
  customer?: Customer;
  conversation_id: string;
  user_id?: number;
  user?: User;
  file_ids?: Array<string>;
  attachments?: Array<Attachment>;
  mentioned_user_ids?: Array<number>;
  mentions?: Array<any>;
  metadata?: any;
};
// ...
export type ConversationSource = 'chat' | 'email' | 'slack' | 'sms';

export type Conversation = {
  id: string;
  source?: ConversationSource;
  subject?: string;
  account_id: string;
  customer_id: string;
  customer: Customer;
  created_at: string;
  updated_at: string;
  closed_at?: string | null;
  archived_at?: string | null;
  last_activity_at?: string;
  date: string;
  preview: string;
  messages?: Array<Message>;
  priority?: string;
  read?: boolean;
  status?: string;
  assignee_id?: number;
  inbox_id?: string;
  tags?: Array<Tag>;
  mentions?: Array<any>;
};

export type CustomerNote = {
  id: string;
  body: string;
  customer_id: string;
  author_id: number;
  created_at: string;
  updated_at: string;
  author?: User;
  customer?: Customer;
};

export type Tag = {
  id: string;
  name: string;
  description?: string;
  color?: string;
  updated_at: string;
};